fix: board 디자인 및 queryKey 수정
