+++ src/app/community/page.tsx
@@ -4,10 +4,10 @@ import type { NextPage } from 'next';
 import Link from 'next/link';
 import FloatingButton from '@components/floating-button';
 import Layout from '@components/layout';
-import { useBoardListQuery } from '@libs/hooks/services/queries/board';
 import React, { useEffect, useState } from 'react';
 import ReactPaginate from 'react-paginate';
 import { ChatBubbleIcon, CheckCircledIcon, Pencil1Icon } from '@radix-ui/react-icons';
+import { useBoardListQuery } from '@/community/queries';
 
 const Community: NextPage = () => {
   const [page, setPage] = useState(0);
@@ -36,37 +36,36 @@ const Community: NextPage = () => {
         </div>
       )}
 
-      <div className="space-y-4 divide-y-[2px]">
+      <div className="divide-y-[2px]">
         {data.list.map((content: IBoardContent) => {
           const { boardId, title, name, createDate } = content;
           return (
             <Link
               key={boardId}
               href={`/community/${boardId}`}
-              className="flex cursor-pointer flex-col items-start pt-4"
+              className="flex cursor-pointer flex-col items-start p-4"
             >
-              <div className="flex">
+              <div className="mt-2 flex">
                 {content.hashTags.map((hashTag: IHashTag, index) => {
                   const { tag } = hashTag;
                   return (
                     <span
                       key={index}
-                      className="rounded-full bg-secondary px-1.5 py-0.5 text-xs font-medium text-gray-800 first:ml-4"
+                      className="mr-1 rounded-lg bg-secondary px-1.5 py-0.5 text-xs font-medium text-muted-foreground"
                     >
                       {tag}
                     </span>
                   );
                 })}
+                <img src="" alt="" />
               </div>
 
-              <div className="mt-2 px-4">
-                <span className="font-medium text-orange-500">Q.</span> {title}
-              </div>
-              <div className="mt-5 flex w-full items-center justify-between px-4 text-xs font-medium text-gray-500">
+              <div className="mt-2">{title}</div>
+              <div className="mt-5 flex w-full items-center justify-between text-xs font-medium">
                 <span>{name}</span>
-                <span>{createDate}</span>
+                <span className="text-muted-foreground">{createDate}</span>
               </div>
-              <div className="mt-3 flex w-full space-x-5 border-t px-4 py-2.5   text-gray-700">
+              <div className="mt-3 flex w-full justify-end space-x-5">
                 <span className="flex items-center space-x-2 text-sm">
                   <CheckCircledIcon className="h-4 w-4" />
                   <span>공감 1</span>
